requirements.txt
openai==0.28
pyodbc==5.1.0
azure-functions==1.18.0
azure-ai-textanalytics==5.3.0
azure-core==1.30.2
azure-cosmos==4.7.0
Flask==3.0.3
thefuzz==0.22.1

